@@ -5,7 +5,7 @@
   "scripts": {
     "ng": "ng",
     "build": "npm run env -s && ng build --prod",
-    "start": "ng serve --proxy-config proxy.conf.js",
+    "start": "ng serve --aot --proxy-config proxy.conf.js",
     "lint": "ng lint --type-check && stylelint \"src/**/*.scss\" --syntax scss && htmlhint \"src\" --config .htmlhintrc",
     "test": "ng test",
     "test:ci": "npm run lint -s && ng test --single-run --code-coverage",
@@ -17,60 +17,59 @@
     "release": "standard-version && git push --follow-tags origin develop"
   },
   "dependencies": {
-    "@angular/animations": "^4.2.6",
-    "@angular/common": "^4.2.6",
-    "@angular/compiler": "^4.2.6",
-    "@angular/core": "^4.2.6",
-    "@angular/forms": "^4.2.6",
-    "@angular/http": "^4.2.6",
-    "@angular/platform-browser": "^4.2.6",
-    "@angular/platform-browser-dynamic": "^4.2.6",
-    "@angular/router": "^4.2.6",
+    "@angular/animations": "^5.0.4",
+    "@angular/common": "^5.0.4",
+    "@angular/compiler": "^5.0.4",
+    "@angular/core": "^5.0.4",
+    "@angular/forms": "^5.0.4",
+    "@angular/http": "^5.0.4",
+    "@angular/platform-browser": "^5.0.4",
+    "@angular/platform-browser-dynamic": "^5.0.4",
+    "@angular/router": "^5.0.4",
+    "@ngx-translate/core": "^9.0.1",
     "@angular/service-worker": "^1.0.0-beta.16",
-    "@ng-bootstrap/ng-bootstrap": "1.0.0-beta.5",
+    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.8",
     "@ngrx/effects": "^4.1.1",
     "@ngrx/store": "^4.1.1",
     "@ngrx/store-devtools": "^4.1.1",
-    "@ngx-translate/core": "^8.0.0",
-    "bootstrap": "4.0.0-beta",
-    "core-js": "^2.4.1",
+    "bootstrap": "4.0.0-beta.2",
     "font-awesome": "^4.7.0",
+    "core-js": "^2.5.3",
     "lodash": "^4.17.4",
-    "ng2-file-upload": "^1.2.1",
-    "ngx-chips": "~1.5.11",
-    "rxjs": "^5.4.0",
-    "zone.js": "^0.8.18"
+    "ngx-chips": "~1.6.2",
+    "rxjs": "5.5.2",
+    "zone.js": "^0.8.14"
   },
   "devDependencies": {
-    "@angular/cli": "^1.4.9",
-    "@angular/compiler-cli": "^4.2.6",
-    "@angular/language-service": "^4.2.6",
-    "@biesbjerg/ngx-translate-extract": "^2.3.4",
-    "@ngx-rocket/scripts": "^1.0.5",
+    "@angular/cli": "~1.5.0",
+    "@angular/compiler-cli": "^5.0.4",
+    "@angular/language-service": "^5.0.4",
+    "@biesbjerg/ngx-translate-extract": "^2.3.1",
+    "@ngx-rocket/scripts": "^1.1.0",
     "@types/jasmine": "^2.5.52",
-    "@types/jasminewd2": "^2.0.3",
-    "@types/lodash": "^4.14.85",
-    "@types/node": "^8.0.51",
-    "codelyzer": "~3.2.2",
+    "@types/jasminewd2": "^2.0.2",
+    "@types/lodash": "^4.14.52",
+    "@types/node": "^8.0.10",
+    "codelyzer": "~4.0.1",
     "hads": "^1.3.4",
     "htmlhint": "^0.9.13",
     "https-proxy-agent": "^2.0.0",
     "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "~1.7.1",
+    "jasmine-spec-reporter": "^4.1.0",
+    "karma": "~1.7.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.0",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "^1.1.0",
     "karma-junit-reporter": "^1.2.0",
     "protractor": "~5.2.0",
     "puppeteer": "^0.13.0",
     "standard-version": "^4.2.0",
-    "stylelint": "~8.4.0",
+    "stylelint": "~8.3.1",
     "stylelint-config-recommended-scss": "~2.0.0",
-    "stylelint-config-standard": "~17.0.0",
-    "stylelint-scss": "~2.1.0",
-    "ts-node": "^3.3.0",
+    "stylelint-config-standard": "~18.0.0",
+    "stylelint-scss": "~2.2.0",
+    "ts-node": "~3.3.0",
     "tslint": "~5.8.0",
     "typescript": "~2.4.2"
   }
