@@ -1,9 +1,8 @@
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-
+import { Observable } from 'rxjs/Observable';
+import { of } from 'rxjs/observable/of';
+import { map, catchError } from 'rxjs/operators';
 import { Injectable } from '@angular/core';
 import { Http, Response, RequestOptions, Headers } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
 import Poll from './../models/poll';
 import News from './../models/news';
 import { AuthenticationService } from '../core/authentication/authentication.service';
@@ -25,16 +24,18 @@ export class NewsService {
 
     getNews(context: NewsContext): Observable<News> {
         return this.http.get(routes.news(context))
-            .map((res: Response) => res.json())
-            .map(body => body)
-            .catch(() => Observable.of('Error, could not news.'));
+            .pipe(
+            map((res: Response) => res.json()),
+            catchError(() => of('Error, no se encontró la noticia.'))
+            );
     }
 
     getAllNews(): Observable<News[]> {
         return this.http.get(routes.allNews())
-            .map((res: Response) => res.json())
-            .map(body => body)
-            .catch(() => Observable.of('Error, could not news.'));
+            .pipe(
+            map((res: Response) => res.json()),
+            catchError(() => of('Error, no se encontraron noticias.'))
+            );
     }
 
 }
