@@ -18,8 +18,10 @@ export class Route {
       path: '',
       component: ShellComponent,
       children: routes,
-      canActivate: []
-    }];
+      canActivate: [],
+      // Reuse ShellComponent instance when navigating between child views
+      data: { reuse: true }
+}];
   }
 
 }
